Move Generic_handle_map.h -> Void_handle_hash_function.h
