+++ Nef_3/include/CGAL/Nef_3/SNC_sphere_map.h
@@ -21,7 +21,6 @@
 #include <CGAL/Unique_hash_map.h>
 #include <CGAL/Nef_2/Object_handle.h>
 #include <CGAL/Nef_S2/SM_iteration.h>
-#include <CGAL/Nef_S2/Generic_handle_map.h>
 #include <CGAL/Nef_2/iterator_tools.h>
 #include <CGAL/Nef_3/Infimaximal_box.h>
 #include <CGAL/Nef_S2/Sphere_geometry.h>
