+++ Nef_S2/include/CGAL/Nef_S2/Void_handle_hash_function.h
